pyproject.toml
[project]
name = "autowz"
version = "0.1.0"
description = "《今天怎么看》微信公众号自动发布系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "jinja2>=3.1.4",
  "markdown>=3.7",
  "python-multipart>=0.0.9",
  "openai>=1.30.0",
  "Pillow>=10.0.0",
  "feedparser>=6.0.0",
]

[project.optional-dependencies]
worker = [
  "apscheduler>=3.10.0",
  "redis>=5.0.0",
]
db = [
  "sqlalchemy>=2.0.0",
  "pymysql>=1.1.0",
]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
  "respx>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
